src/requirements.txt
# https://github.com/pyeve/cerberus
cerberus~=1.3

# https://github.com/dateutil/dateutil
python-dateutil~=2.8